requirements.txt
streamlit
llama-index
llama-index-vector-stores-chroma
llama-index-embeddings-openai
llama-index-llms-openai
chromadb
python-dotenv
tqdm
PyMuPDF
markdown